chore: 创建统一的拖拽预览
